More hygeinic (and performant?) colon detection.
This change leverages the Token linked list to split rules in a more
hygeinic way. Previously, we were slicing the string and then using
`indexOf` to find the `colon` and then splitting the string on its
index. Now we can just traverse the Token linked list and compare Token
instances to do the same thing.